+++ index.js
@@ -1,10 +1,172 @@
 'use strict';
-module.exports =  function(config) {
-    config.tz = config.tz || 'Europe/Amsterdam';
-    config.locale = config.locale || 'nl';
-    return {
-        templateDataFactory: require('./src/templateDataFactory.js')(config),
-        handlebarsHelpers: require('./src/handlebarsHelpers.js')(config),
-        renderFile: require('./src/file.js')(config)
-    };
+var _ = require('lodash'),
+	when = require('when');
+
+module.exports = function (config) {
+	var cbc = config.cbc || require('communibase-connector-js');
+	var maxNestLevel = config.maxNestLevel || 5;
+
+	var entitiesHashPromise = cbc.getAll('EntityType').then(function (entities) {
+		var entitiesHash = {};
+		_.each(entities, function (entity) {
+			entitiesHash[entity.title] = entity;
+		});
+		return entitiesHash;
+	});
+
+
+	/**
+	 * Returns all data to be assign to template, based on the wizard-source object
+	 * @param {String} entityTypeTitle - The entity type (i.e. Person, Shift, etc)
+	 * @param {Object} document - The actual document (entity) instance which should be templatified
+	 * @param {Number} [nestLevel] - Optional current nesting-level to prevent infinite loops. Requests will queue
+	 * if nesting level is undefined to fix https://trello.com/c/Ia8QR75d/652-opstellen-herinneringen-blijft-hangen-op-
+	 * @returns {Object}
+	 */
+	this.getPromise = function (entityTypeTitle, document, nestLevel) {
+		var result = { _id: document._id }, subPromises = [], self = this;
+
+		if (nestLevel === undefined) {
+			nestLevel = 0;
+		}
+
+		if (nestLevel > maxNestLevel) {
+			return when({});
+		}
+
+		return entitiesHashPromise.then(function (entitiesHash) {
+			var entityType = entitiesHash[entityTypeTitle];
+
+			//Expose _ALL_ attributes (not just commmunibase fields): https://trello.com/c/9yKbd7Zg/460-wat-klopt-er-nie
+			_.each(entityType.attributes, function (attribute) {
+				var value = document[attribute.title], arrayItemPromises;
+
+				if (_.contains([undefined, null, true, false], value) || _.isDate(value) || _.isNumber(value)) {
+					result[attribute.title] = value;
+					return;
+				}
+
+				if (_.isArray(value)) {
+					if (attribute.ref) {
+						result[attribute.title] = attribute.ref;
+
+						if (!entitiesHash[attribute.ref] || (attribute.title.substr(-3) !== 'Ids')) {
+							return;
+						}
+
+						//applicableForGroupIds => applicableForGroups
+						result[attribute.title.substr(-3) + 's'] = [];
+						subPromises.push(cbc.getByIds(attribute.ref, value).then(function (subDocuments) {
+							var subSubPromises = [];
+							_.each(subDocuments, function (subDocument) {
+								subSubPromises.push(
+									self.getPromise(attribute.ref, subDocument, (nestLevel + 1)).then(
+											function (templateData) {
+												result[attribute.title.substr(-3) + 's'].push(templateData);
+											}
+									)
+								);
+							});
+							return when.all(subSubPromises);
+						}));
+						return;
+					}
+
+					result[attribute.title] = [];
+					//We need to maintain the array order --> process and add them in order of the original array!
+					arrayItemPromises = [];
+					_.each(value, function (subDocument) {
+						//check for strings, e.g. Vasmo Company.classifications
+
+						if (subDocument && entitiesHash[attribute.items]) {
+							arrayItemPromises.push(self.getPromise(attribute.items, subDocument,
+									(nestLevel + 1)));
+							return;
+						}
+						//Not a document? push the raw value (string)
+						arrayItemPromises.push(value);
+					});
+
+					//We need to maintain the array order --> process and add them in order of the original array!
+					subPromises.push(when.all(arrayItemPromises).then(function (templateDatas) {
+						result[attribute.title] = templateDatas;
+					}));
+					return;
+				}
+
+				if (entitiesHash[attribute.type]) {
+					if (attribute.title.substr(-9) !==  'Reference') {
+						subPromises.push(self.getPromise(attribute.type, value, (nestLevel + 1)).then(
+							function (templateData) {
+								result[attribute.title] = templateData;
+							}
+						));
+						return;
+					}
+
+					var referredDocumentProperty = attribute.title.substr(0, (attribute.title.length - 9));
+
+					if (attribute.type ===  'DocumentReference') {
+						subPromises.push(
+							cbc.getByRef(value, document).then(
+								function (referredDocument) {
+									result[referredDocumentProperty] = referredDocument;
+								}, function () {
+									result[referredDocumentProperty] = null;
+									return when();
+								}
+							)
+						);
+						return;
+					}
+
+					//something like membership.emailAddressReference
+					if (value.documentReference) {
+						subPromises.push(
+							cbc.getByRef(value.documentReference, value).then(
+								function (referredDocument) {
+									result[referredDocumentProperty] = referredDocument;
+								}, function () {
+									result[referredDocumentProperty] = null;
+									return when();
+								}
+							)
+						);
+						return;
+					}
+
+					// a custom defined address / phoneNumber / emailAddress within a reference
+					result[referredDocumentProperty] = value[referredDocumentProperty];
+				}
+
+				// Default key-value store
+				result[attribute.title] = value;
+				if (value === '') {
+					return;
+				}
+
+				// A special ObjectId-string? --- enrich the template data!
+				if (attribute.type === 'ObjectId' &&
+						attribute.ref &&
+						entitiesHash[attribute.ref] &&
+						entitiesHash[attribute.ref].isResource &&
+						attribute.title.substr(-2) === 'Id') {
+
+					subPromises.push(cbc.getById(attribute.ref, value).then(function (referredObject) {
+						return self.getPromise(attribute.ref, referredObject, nestLevel + 1).then(
+							function (templateData) {
+								result[attribute.title.substr(0, (attribute.title.length - 2))] = templateData;
+							}
+						);
+					}, function (/*err*/) {
+						return when.resolve({});
+					}));
+				}
+			});
+
+			return when.all(subPromises).then(function () {
+				return result;
+			});
+		});
+	};
 };
