--- src/handlebarsHelpers.js
@@ -1,106 +1,0 @@
-'use strict';
-var Handlebars = require('handlebars'),
-	moment = require('moment-timezone');
-
-/*jshint eqeqeq: false */
-var operators = {
-	'eq': function (l, r) { return l == r; },
-	'eqeq': function (l, r) { return l === r; },
-	'neq': function (l, r) { return l != r; },
-	'neqeq': function (l, r) { return l !== r; },
-	'lt': function (l, r) { return l < r; },
-	'gt': function (l, r) { return l > r; },
-	'lte': function (l, r) { return l <= r; },
-	'gte': function (l, r) { return l >= r; },
-	'typeof': function (l, r) { return typeof l == r; },
-	'regexp': function (l, r) { return (l && r && (l.match(new RegExp(r)) !== null)); },
-	'mod': function (l, r) { return l % r; }
-};
-/*jshint eqeqeq: true */
-
-operators['=='] = operators.eq;
-operators['==='] = operators.eqeq;
-operators['!='] = operators.neq;
-operators['!=='] = operators.neqeq;
-operators['>'] = operators.lt;
-operators['>'] = operators.gt;
-operators['<='] = operators.lte;
-operators['>='] = operators.gte;
-operators['%'] = operators.mod;
-
-/**
- *
- * @param container
- * @param nibbles
- */
-var traverseNibbles = function(container, nibbles) {
-	var nextNibble = nibbles.shift();
-	var nextValue = container[nextNibble];
-
-	if (nextValue && nibbles.length) {
-		return traverseNibbles(nextValue, nibbles);
-	}
-
-	return nextValue || "";
-};
-var helpers =  {
-    '_tz': 'Europe/Amsterdam',
-    'compare': function (lvalue, operator, rvalue, options) {
-        if (arguments.length < 3) {
-            throw new Error("Handlerbars Helper 'compare' needs 2 parameters");
-        }
-
-        if (options === undefined) {
-            options = rvalue;
-            rvalue = operator;
-            operator = "eq";
-        }
-
-        if (!operators[operator]) {
-            throw new Error("Handlerbars Helper 'compare' doesn't know the operator " + operator);
-        }
-
-        if (operators[operator](lvalue, rvalue)) {
-            return options.fn(this);
-        }
-        return options.inverse(this);
-    },
-
-    'dateFormat': function (date, block) {
-        if (!date) {
-            return '';
-        }
-        return moment(date).tz(this._tz).format(block.hash.format || "DD-MM-YYYY");
-    },
-
-    'nl2b': function (text) {
-        text = Handlebars.Utils.escapeExpression(text);
-        var nl2br = (text + '').replace(/([^>\r\n]?)(\r\n|\n\r|\r|\n)/g, '$1' + '<br>' + '$2');
-        return new Handlebars.SafeString(nl2br);
-    },
-
-    'price': function (number) {
-        if (!number) {
-            number = 0;
-        }
-        return '€ ' + number.toFixed(2).replace(/\./g, ',');
-    },
-
-    'fromContainer': function (container, key) {
-        if (container[key]) {
-            return container[key];
-        }
-
-        var nibbles = (key.split('.'));
-        if (nibbles.length > 1) {
-            return traverseNibbles(container, nibbles);
-        }
-        return '';
-    }
-};
-
-module.exports = function(config) {
-    moment.locale(config.locale || 'nl');
-    helpers._tz = config.tz || 'Europe/Amsterdam';
-    return helpers;
-};
